removed emitter library
